Important FIX for bug in metadata report whereby the original file is DELETED if the attempted metadata write fails

+++ pages/ajax/metadata_report.php
@@ -54,12 +54,12 @@ else
 
     # Create a temporary file (simulate download) and create a report for it.
     $tmpfile = write_metadata($image, $ref);
-    if ($tmpfile===false || !file_exists($tmpfile)){$tmpfile=$image;}
+    if ($tmpfile===false || !file_exists($tmpfile)){$tmpfile=$image;$donotdelete=true;}
     $command = $exiftool_fullpath . " -s -t -G --filename --exiftoolversion --filepermissions --NativeDigest --History --Directory " . escapeshellarg($tmpfile)." 2>&1";
     $report_simulated = run_command($command);
 
     # Remove the temporary file.
-    unlink($tmpfile);
+    if (!isset($donotdelete)){unlink($tmpfile);}
 
     # Process the report of the simulated download.
     $results_simulated = array();
